+++ CHANGELOG.md
@@ -13,9 +13,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * system_param
   * display
   * ctrl
+  * kernel::rng
 
 ### Fixed
 - Fix `i32` being used instead of `u32` as `psp2_sys::types::SceUInt`
+
+### Changed
+- Type of variable used by `size` parameter in `kernel::rng::sceKernelGetRandomNumber` function. (`SceSize` instead of `u32`)
+
 
 [Control Library]: https://docs.vitasdk.org/group__SceCtrlUser.html